+++ version.sbt
@@ -1,1 +1,1 @@
-version in ThisBuild := "0.13.1-SNAPSHOT"
+version in ThisBuild := "0.14.0-SNAPSHOT"
